requirements.txt
google-cloud-bigquery>=3.27.0
google-auth>=2.37.0
tqdm>=4.67.1
pinecone-client==5.0.1
numpy==2.2.6